+++ demisto_sdk/commands/content_graph/objects/integration.py
@@ -1,11 +1,12 @@
-from typing import TYPE_CHECKING, List, Set
+from typing import TYPE_CHECKING, List
+
+from demisto_sdk.commands.content_graph.objects.base_content import BaseContent
 
 if TYPE_CHECKING:
     # avoid circular imports
     from demisto_sdk.commands.content_graph.objects.script import Script
-    from demisto_sdk.commands.content_graph.objects.relationship import RelationshipData
 
-from pydantic import BaseModel, Field, validator
+from pydantic import Field
 
 from demisto_sdk.commands.content_graph.common import (ContentType,
                                                        RelationshipType)
@@ -13,20 +14,27 @@ from demisto_sdk.commands.content_graph.objects.integration_script import \
     IntegrationScript
 
 
-class Command(BaseModel):
+class Command(BaseContent, content_type=ContentType.COMMAND):  # type: ignore[call-arg]
     name: str
-    object_id: str = ""  # objects sets up in the validator
+
+    # From HAS_COMMAND relationship
     deprecated: bool = False
     description: str = ""
 
-    relationships_data: Set["RelationshipData"] = Field(set(), exclude=True, repr=False)  # too much data in the repr
+    # missing attributes in DB
+    node_id: str = ""
+    object_id: str = Field("", alias="id")
 
-    @validator("object_id", always=True)
-    def validate_object_id(cls, v, values):
-        return values["name"]
+    @property
+    def integrations(self) -> List["Integration"]:
+        return [
+            r.content_item  # type: ignore[misc]
+            for r in self.relationships_data[RelationshipType.HAS_COMMAND]
+            if r.content_item == r.source
+        ]
 
-    class Config:
-        orm_mode = True
+    def dump(self, *args) -> None:
+        raise NotImplementedError()
 
 
 class Integration(IntegrationScript, content_type=ContentType.INTEGRATION):  # type: ignore[call-arg]
@@ -39,21 +47,21 @@ class Integration(IntegrationScript, content_type=ContentType.INTEGRATION):  # t
     @property
     def imports(self) -> List["Script"]:
         return [
-            r.content_item
-            for r in self.relationships_data
-            if r.relationship_type == RelationshipType.IMPORTS and r.content_item == r.target
+            r.content_item  # type: ignore[misc]
+            for r in self.relationships_data[RelationshipType.IMPORTS]
+            if r.content_item == r.target
         ]
 
     def set_commands(self):
         commands = [
             Command(
                 # the related to has to be a command
-                name=r.content_item.name,  # type: ignore[union-attr]
+                name=r.content_item.name,  # type: ignore[union-attr,attr-defined]
+                marketplaces=self.marketplaces,
                 deprecated=r.deprecated,
                 description=r.description,
             )
-            for r in self.relationships_data
-            if r.is_direct and r.relationship_type == RelationshipType.HAS_COMMAND
+            for r in self.relationships_data[RelationshipType.HAS_COMMAND]
         ]
         self.commands = commands