+++ demisto_sdk/commands/content_graph/objects/pack.py
@@ -4,6 +4,7 @@ from collections import defaultdict
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Generator, List, Optional
 
+from packaging.version import parse
 from pydantic import BaseModel, Field
 
 from demisto_sdk.commands.common.constants import (
@@ -154,22 +155,45 @@ class Pack(BaseContent, PackMetadata, content_type=ContentType.PACK):  # type: i
 
     @property
     def depends_on(self) -> List["RelationshipData"]:
+        """
+        This returns the packs which this content item depends on.
+        In addition, we can tell if it's a mandatorily dependency or not.
+
+        Returns:
+            List[RelationshipData]:
+                RelationshipData:
+                    relationship_type: RelationshipType
+                    source: BaseContent
+                    target: BaseContent
+
+                    # this is the attribute we're interested in when querying
+                    content_item: BaseContent
+
+                    # Whether the relationship between items is direct or not
+                    is_direct: bool
+
+                    # Whether using the command mandatorily (or optional)
+                    mandatorily: bool = False
+
+        """
         return [
             r
-            for r in self.relationships_data
-            if r.relationship_type == RelationshipType.DEPENDS_ON and r.content_item == r.target
+            for r in self.relationships_data[RelationshipType.DEPENDS_ON]
+            if r.content_item == r.target
         ]
 
     def set_content_items(self):
-        content_items = [
-            r.content_item
-            for r in self.relationships_data
-            if r.relationship_type == RelationshipType.IN_PACK and r.content_item == r.source
+        content_items: List[ContentItem] = [
+            r.content_item  # type: ignore[misc]
+            for r in self.relationships_data[RelationshipType.IN_PACK]
+            if r.content_item == r.source
         ]
         content_item_dct = defaultdict(list)
         for c in content_items:
             content_item_dct[c.content_type].append(c)
 
+        # If there is no server_min_version, set it to the maximum of its content items fromversion
+        self.server_min_version = self.server_min_version or str(max(parse(content_item.fromversion) for content_item in content_items))
         self.content_items = PackContentItems.parse_obj(content_item_dct)
 
     def dump_metadata(self, path: Path) -> None: