+++ demisto_sdk/commands/content_graph/objects/playbook.py
@@ -5,7 +5,6 @@ from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
 
 
 class Playbook(ContentItem, content_type=ContentType.PLAYBOOK):  # type: ignore[call-arg]
-    is_test: bool
 
     def metadata_fields(self) -> Set[str]:
         return {"name", "description"}