+++ demisto_sdk/commands/content_graph/parsers/widget.py
@@ -17,6 +17,10 @@ class WidgetParser(JSONContentItemParser, content_type=ContentType.WIDGET):
 
         self.connect_to_dependencies()
 
+    @property
+    def marketplaces(self) -> List[MarketplaceVersions]:
+        return [MarketplaceVersions.XSOAR]
+
     def connect_to_dependencies(self) -> None:
         """Collects the playbook used in the widget as a mandatory dependency."""
         if self.data_type == "scripts":