+++ demisto_sdk/commands/content_graph/parsers/xsiam_report.py
@@ -21,3 +21,7 @@ class XSIAMReportParser(JSONContentItemParser, content_type=ContentType.XSIAM_RE
     @property
     def object_id(self) -> Optional[str]:
         return self.json_data.get("global_id")
+
+    @property
+    def marketplaces(self) -> List[MarketplaceVersions]:
+        return [MarketplaceVersions.MarketplaceV2]